refactor(i18n): Update routing imports to use new i18n structure across authentication and main components
