+++ src/components/dashboard-aside/DashboardAside.tsx
@@ -16,7 +16,7 @@ import {
   TooltipTrigger,
   TooltipProvider,
 } from "@/components/ui/tooltip";
-import { Link } from "@/routes";
+import { Link } from "@/i18n/routes";
 
 export default function DashboardAside() {
   const t = useTranslations("Navigation");
